pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "app"
version = "0.1.0"
description = "Basic Flask app for agentic AI it response with a html of analysis"
requires-python=">=3.12.0"

dependencies = [
  "flask>=3.0.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "playwright>=1.40.0",
  "psycopg>=3.1.0",
  "psycopg_pool>=3.2.0",
  "langchain==0.3.25",
  "langchain-core==0.3.64",
  "langchain-community==0.3.24",
  "langchain-experimental==0.3.4",
  "langchain-text-splitters==0.3.8",
  "langchain-openai==0.3.21",
  "langchain-anthropic==0.3.15",
  "langgraph==0.4.8",
  "langgraph-checkpoint==2.0.26",
  "langgraph-checkpoint-sqlite==2.0.10",
  "langgraph-prebuilt==0.2.2",
  "langgraph-sdk==0.1.70",
  "langsmith==0.3.45",
  "lxml>=4.9",
  "mcp>=1.0.0",
  "langchain-mcp-adapters==0.1.14",
  "flask[async]",
  "beautifulsoup4",
  "gunicorn" 
]